Attempted to fix issue on older iOS devices where the blur is too bright

+++ Assets/Standard Assets/Effects/ImageEffects/Shaders/_BloomAndFlares/MobileBlur.shader
@@ -47,7 +47,8 @@ Shader "Hidden/FastBlur" {
 		
 		fixed4 fragDownsample ( v2f_tap i ) : SV_Target
 		{				
-			fixed4 color = tex2D (_MainTex, i.uv20);
+			//fixed4 color = tex2D (_MainTex, i.uv20);
+			half4 color = tex2D (_MainTex, i.uv20);
 			color += tex2D (_MainTex, i.uv21);
 			color += tex2D (_MainTex, i.uv22);
 			color += tex2D (_MainTex, i.uv23);
